pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "reevo"
version = "0.1.0"
description = "ReEvo: Large Language Models as Hyper-Heuristics with Reflective Evolution"
authors = [
    { name = "Haoran Ye", email = "hrye@stu.pku.edu.cn"},
]
requires-python = ">=3.11"
readme = "README.md"
license = "MIT"
dependencies = [
    "annotated-types>=0.6.0",
    "antlr4-python3-runtime>=4.9.3",
    "anyio>=4.2.0",
    "certifi>=2024.7.4",
    "distro>=1.9.0",
    "h11>=0.14.0",
    "httpcore>=1.0.2",
    "httpx>=0.26.0",
    "hydra-core>=1.3.2",
    "idna>=3.7",
    "numpy>=1.26.3",
    "omegaconf>=2.3.0",
    "openai>=1.8.0",
    "packaging>=23.2",
    "pydantic>=2.5.3",
    "pydantic_core>=2.14.6",
    "PyYAML>=6.0.1",
    "scipy>=1.11.4",
    "sniffio>=1.3.0",
    "tqdm>=4.66.3",
    "typing_extensions>=4.9.0",
    "pip>=24.0.1",
    "hydra-colorlog>=1.2.0",
]

[project.optional-dependencies]
aco = ["torch", "scikit-learn"]
gls = ["numba>=0.58.0"]
nco = ["torch"]
dev = ["black", "pre-commit", "ruff"]

[tool.ruff]
select = ["F", "E", "W", "I001"]
line-length = 90
show-fixes = false
target-version = "py311"
task-tags = ["TODO", "FIXME"]
ignore = ["E501"] # never enforce `E501` (line length violations), done in Black

[tool.ruff.per-file-ignores]
"__init__.py" = ["E402", "F401"]

[tool.ruff.isort]
known-first-party = []
known-third-party = []
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder",
]
combine-as-imports = true
split-on-trailing-comma = false
lines-between-types = 1

[tool.black]
line-length = 90
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
(
  /(
      \.direnv
    | \.eggs
    | \.git
    | \.tox
    | \.venv
    | _build
    | build
    | dist
    | venv
  )/
)
'''

[tool.coverage.report]
show_missing = true
exclude_lines = [
    # Lines to exclude from coverage report (e.g., comments, debug statements)
    "pragma: no cover",
    "if **name** == .__main__.:",
]
